+++ lib/data/sources/repositories/compose_repository_impl.dart
@@ -13,7 +13,7 @@ class ComposeRepositoryImpl extends ComposeRepository {
 
   @override
   Future<RawTransaction> composeSend(
-      String sourceAddress, String destination, String asset, int quantity,
+      String sourceAddress, String destination, String asset, double quantity,
       [bool? allowUnconfirmedTx, int? fee]) async {
     final response = await api.composeSend(
         sourceAddress, destination, asset, quantity, allowUnconfirmedTx, fee);
@@ -27,7 +27,7 @@ class ComposeRepositoryImpl extends ComposeRepository {
 
   @override
   Future<compose_send.ComposeSend> composeSendVerbose(
-      String sourceAddress, String destination, String asset, int quantity,
+      String sourceAddress, String destination, String asset, double quantity,
       [bool? allowUnconfirmedTx, int? fee]) async {
     final response = await api.composeSendVerbose(
         sourceAddress, destination, asset, quantity, allowUnconfirmedTx, fee);
